[tests] Add test for apk size comparison
Context: https://github.com/xamarin/xamarin-android/issues/5214

Build simple XA and XA/Forms apps in tests with current and NET5/6 XA.
Do not remove the temporary projects, so that it is possible to compare
the apks after tests run.

The apps are built with single arm64 ABI, in Release configuration.

The test can be run like this.

current XA:

    msbuild Xamarin.Android.sln /t:RunNunitTests /p:TEST="Xamarin.Android.Build.Tests.BuildTest.BuildReleaseArm64"

NET5/6:

    dotnet test -v diag --filter BuildTest.BuildReleaseArm64 .\bin\TestDebug\netcoreapp3.1\Xamarin.Android.Build.Tests.dll