Remove Microsoft.Android.Sdk.DefaultItems.props
The `@(FrameworkReference)` was breaking builds for netstandard2.0
libraries:

    Microsoft.NET.Sdk.FrameworkReferenceResolution.targets(324,5): error NETSDK1073:
    The FrameworkReference 'Microsoft.Android' was not recognized

I moved the `@(FrameworkReference)` to `Microsoft.Android.Sdk.targets`.

I also completely removed `Microsoft.Android.Sdk.DefaultItems.props`
and put the default include wildcards directly in `AutoImport.props`.
I think this will be clearer that this is a special file to be treated
with care.