ExtractWorkloadPacks should <Touch/> all files
It looked to me like the extracted files had UTC timestamps on my
local machine.

This caused various MSBuild tests to fail for me, because all the
`.targets` files' timestamps are in the future. In one example, the
`CoreCompile` target was running every time.

I changed this to `<Touch/>` every extracted file, so this won't be an
issue.