[tests] use latest Xamarin.GooglePlayServices in tests (#5213)
* Bump to Xamarin.GooglePlayServices 117.x.

Running under a `dotnet` context, a MSBuild tests using Google Play
Services would fail because of a `<ProcessGoogleServicesJson/>`
MSBuild task. The assembly was built for .NET framework, and so it
fails when running with `dotnet build`. We can update to
Xamarin.GooglePlayServices 117.x, because it is a netstandard2.0
assembly now.

I did some other general cleanup here:

* Setup `XamarinFormsMapsApplicationProject` so it will bring in the
  newest Xamarin.GooglePlayServices.
* Removed old Xamarin.GooglePlayServices packages from `KnownPackages`.
* Updated tests to use `XamarinFormsMapsApplicationProject` instead of
  the various GooglePlayServices packages directly.
* Removed the relevant `[Category ("DotNetIgnore")]`.

I also completely removed `AndroidUpdateResourcesTest.Issue2205`,
because it seems completely duplicate of
`BuildTest.BuildXamarinFormsMapsApplication` and other tests.
`Issue2205` was only checking that build succeeds for a project with a
`@(GoogleServicesJson)` item, and `XamarinFormsMapsApplicationProject`
has a `@(GoogleServicesJson)` file.