Remove Preserve attribute from IntentService ctor
It shouldn't be needed as the service based on IntentService should
have the constructor without parameters and thus it would preserve its
base constructor.

In case someone declares Android service without public default
parameterless constructor, the build fails with:

    error XA4213: The type 'MyService' must provide a public default constructor

Context: The constructor is called when the service is created by
Android and it uses the parameterless constructor.