[tests/MSBuildDeviceIntegration] Add test for dll's which have the same size.
Context https://github.com/xamarin/monodroid/issues/1123

There seems to be a bug in the new fast deployment system
where if the dll was modified BUT had exactly the same size
as the existing dll on the device it would not be uploade.

This commit adds a test to make sure this does work as
expected.