[.vscode] Update tasks to request Configuration (#5227)
Within Visual Studio Code, various tasks were hardcoded to build the
`Debug` configuration.  This is annoying if you ever need to build
the `Release` configuration, as VSCode won't help you.

Update the VSCode tasks so that the appropriate build configuration
can be specified when you launch the task.  The default configuration
remains `Debug`.

Additionally, add `streetsidesoftware.code-spell-checker` to the
suggested list of extensions.