[One .NET] temporary Windows & macOS installers for .NET 6 (#5225)
For .NET 6 Preview 1, we will need to provide our own installers for
Android workload support (91498d55).  This will also unblock the
Xamarin.Forms / MAUI team as they build on top of the iOS and Android
workloads for .NET 6.

We will eventually do some kind of "insertion" process to provide our
`.nupkg` files to the dotnet/installer repo, so the installers
generated as part of this commit will go away completely, eventually.

For now, create two new installers:

  * macOS `Microsoft.Android.Workload.pkg`:
    installs to `/usr/local/share/dotnet/`

  * Windows `Microsoft.Android.Workload.msi`:
    installs to `C:\Program Files\dotnet\`

Both installers have the following file structure underneath the root
directory:

  * `sdk\5.0.100-rtm.20509.5\EnableWorkloadResolver.sentinel`
  * `sdk-manifests\5.0.100\Microsoft.Android.Workload\**`
  * `packs\Microsoft.Android.Ref\**`
  * `packs\Microsoft.Android.Sdk\**`

The installers will have a hard dependency on .NET 5.0.100-rtm.20509.5,
so we will need to have clear instructions for installing the correct
version of .NET for the Android workload.

The Windows installer is made using WiX, to mirror what
dotnet/installer is using:

  * https://wixtoolset.org/
  * https://github.com/dotnet/installer/blob/861a1dd12cb80bd834d0e51442d46ee7d1a4023f/src/redist/targets/GenerateMSIs.targets

The `.msi` will need to be built on Windows and the `.pkg` will need
to be built on macOS.  `create-dotnet-msi.csproj` will download the
WIX toolset to `~\android-toolchain\wix` and call `candle.exe` and
`light.exe` appropriately.  `create-dotnet-pkg.csproj` is based on
`create-pkg.csproj`, and no additional tooling is needed.

Changes to CI:

  * The `create-installers` make target now creates the .NET 6 `.pkg`
    installer.  The `mac_build` stage creates one additional installer.

  * Any subsequent stages that download the `installers` artifact, now
    use a wildcard for the installer they need.  This will improve some
    of the test stages that were downloading both the `.vsix` and `.pkg`
    installers before.

  * A `.NET 6 Preview Installers` Github status will appear that
    contains public download links for the new installers.

  * A `Build Results - .NET 6 Preview Installers` artifact will contain
    additional build logs.