[One .NET] <CreateAar/> should generate resource case map
The following situation was broken in .NET 6 after the new `.aar` file
implementation in fcd7cf8:

* A class library includes `Resources\layout\Foo.axml` with an
  uppercase name.
* An application references the library and uses the
  `Resource.designer.cs` definition of `Resource.Layout.Foo`.
* `Resource.Layout.Foo` will have the incorrect integer value and
  crash at runtime!

Reviewing the `Resource.designer.cs` file of the library:

    // aapt resource value: 0x7F030000
    public static int Foo = 2130903040;

Compared to the app:

    // aapt resource value: 0x7F030000
    public const int Foo = 2130903040;
    // aapt resource value: 0x7F030001
    public const int foo = 2130903041;

The reason this works in "legacy" Xamarin.Android is the
`__res_name_case_map.txt` file that is generated in
`__AndroidLibraryProjects__.zip`:

https://github.com/xamarin/xamarin-android/blob/57c5a5fde5efd23f5958cfd8119b7f9c31d9e39d/src/Xamarin.Android.Build.Tasks/Tasks/CreateManagedLibraryResourceArchive.cs#L100-L105

If we generate the same `__res_name_case_map.txt` file and place it in
the `.aar` file we generate, this problem is solved.

We decided to place the file in a `.net` directory, as well as moving
the `.netenv` directory for `@(AndroidEnvironment)` files to `.net/env`:

    .net/__res_name_case_map.txt
    .net/env/MyAndroidEnvironment.txt

Going forward, we could potentially put other files in the `.net` folder.

I updated a test for this scenario; it checks we get the proper casing
for a `MyLayout.axml` in `Resource.designer.cs`.