[One .NET] fix linker output if $(AppendRuntimeIdentifierToOutputPath) is false
Many of our existing MSBuild tests set
`$(AppendRuntimeIdentifierToOutputPath)` to false so that the build
output has the same directory structure for "legacy" Xamarin.Android
as for .NET 6.

This causes a problem with the `ILLink` MSBuild target with multiple
`$(RuntimeIdentifiers)`, as linked .NET assemblies are output to:

    obj\Debug\android.21-arm64\linked\Foo.dll
    obj\Debug\android.21-x86\linked\Foo.dll

`AppendRuntimeIdentifierToOutputPath=false` cause both files to be
written to:

    obj\Debug\linked\Foo.dll
    obj\Debug\linked\Foo.dll

And so the last `$(RuntimeIdentifier)` wins, overwriting the linker
output of any previous RIDs. This would cause runtime crashes for some
ABIs, as `System.Private.CoreLib.dll` is different per architecture.

To fix this, we can always force `$(AppendRuntimeIdentifierToOutputPath)`
to be true in the nested build that runs the `ILLink` MSBuild target:

    <PropertyGroup>
      <_AdditionalProperties>
        AppendRuntimeIdentifierToOutputPath=true;
        _OuterIntermediateOutputPath=$(IntermediateOutputPath);
        <!-- ... -->
      </_AdditionalProperties>
    </PropertyGroup>
    <MSBuild
        Projects="$(MSBuildProjectFile)"
        Targets="_ComputeFilesToPublishForRuntimeIdentifiers"
        Properties="RuntimeIdentifier=%(_RIDs.Identity);$(_AdditionalProperties)">
    <!-- ... -->

The only other property to fix up is `$(_AndroidLinkFlag)` to be based
on `$(_OuterIntermediateOutputPath)`, so that the
`_TouchAndroidLinkFlag` MSBuild target has the correct path. This way
the `Inputs/Outputs` of the `_RemoveRegisterAttribute` MSBuild task
will work correctly, as it needs to do its work on the output of the
`ILLink` MSBuild target.

I updated the `AndroidXMigration` test that verifies that we get a
`obj\Debug\android.21-arm64\linked\Xamarin.Forms.Platform.Android.dll`
assembly with the correct AndroidX types present.