Bump to .NET 6.0.100-alpha.1.20562.2
This version of .NET 6 includes support for
`$(AfterMicrosoftNETSdkTargets)`, and so we now have a slightly
different import ordering:

1. `AutoImport.props`
2. The user's `.csproj` file
3. Some early parts of `Microsoft.NET.Sdk` / `Sdk.targets`
4. `Microsoft.Android.Sdk.targets`
5. The rest of `Microsoft.NET.Sdk` / `Sdk.targets`
6. `Microsoft.Android.Sdk.After.targets`

Anything that needs to happen *after* the `Microsoft.NET.Sdk` targets
can go in `After.targets`.

This allows us to set up some defaults that we lost in 91498d55 when
implementing Android as a .NET "workload":

* `$(GenerateDependencyFile)` defaults to `false`
* `$(ProduceReferenceAssembly)` defaults to `false` for application
  projects.

Using `$(SelfContained)` also triggers new behavior in .NET 6 that we
need to turn off:

* `$(GenerateRuntimeConfigurationFiles)` to `false` or we get a
  `*.runtimeconfig.dev.json` file generated for Android applications.
* `$(CopyLocalLockFileAssemblies)` to `false` - otherwise every `.dll`
  file is copied to `$(OutputPath)`, and we don't need this behavior.
  This appears to be new behavior in .NET 6 when `$(SelfContained)` is
  `true`.
* `$(ComputeNETCoreBuildOutputFiles)` to `false` or we get a `*.exe`
  .NET app host in the output directory.

I also adjusted the naming for `windows-x64` to `win-x64`, as
described in 011a7713.