[One .NET] better logic for deduplication of architecture-specific assemblies
Fixes: https://github.com/xamarin/xamarin-android/issues/5263

In .NET 6, we currently run the `ILLink` MSBuild target for each
`$(RuntimeIdentifier)`, as the Mono runtime packs have some .NET
assemblies that differ per architecture.

To make this happen, the `<ProcessAssemblies/>` MSBuild task works
with input such as:

* `android.21-arm\linked\Foo.dll`
* `android.21-x86\linked\Foo.dll`
* `android.21-arm\linked\System.Private.CoreLib.dll`
* `android.21-x86\linked\System.Private.CoreLib.dll`

And "merges" this to what we need in the output of the `.apk`:

* `assemblies\Foo.dll`
* `assemblies\armeabi-v7a\System.Private.CoreLib.dll`
* `assemblies\x86\System.Private.CoreLib.dll`

`Foo.dll` is identical for all architectures, so we only need to use a
single copy of it.

A bug in our implementation was for an assembly when building for
three RIDs:

* `android.21-arm\linked\System.Collections.Concurrent.dll`
* `android.21-arm64\linked\System.Collections.Concurrent.dll`
* `android.21-x86\linked\System.Collections.Concurrent.dll`

Two of these assemblies have the same MVID... This didn't quite work.

So I reworked the logic to instead do:

* Group by assembly file name
* Count the unique MVIDs
* If we have only 1 MVID, we just need the first assembly.
* Otherwise, place the assemblies in architecture-specific directories.

The resulting code seems like it is a bit simpler, and likely a little
faster. I was able to call `MonoAndroidHelper.HasMonoAndroidReference()`
once per file name, which should help performance.

I updated `XASdkTests` to check the broken scenario.