[One .NET] fix Release builds with AndroidLinkMode=None (#5277)
Fixes: https://github.com/xamarin/xamarin-android/issues/5274

In 51fb93e1, I had logic checking for `$(AndroidLinkMode)` and
`$(AndroidIncludeDebugSymbols)` in two places:

* The `<ProcessAssemblies/>` MSBuild task, which eventually populates
  a `@(_ShrunkAssemblies)` item group.
* The `_PrepareAssemblies` MSBuild target, which fills out various
  item groups such as `@(_ShrunkFrameworkAssemblies)` and
  `@(_ShrunkUserAssemblies)`.

The logic did not match exactly, and so building a `Release` app with
`AndroidLinkMode=None` would result in all of these item groups being
empty.

I switched to checking `$(PublishTrimmed)` as it is the only property
that needs to be checked here:

* if `$(PublishTrimmed)` is `true`, `ILLink` runs and we need to use
  the assemblies that the `_RemoveRegisterAttribute` MSBuild target
  outputs to the `shrunk` directory.
* if `$(PublishTrimmed)` is `false`, we need to use the assemblies
  that` `_LinkAssembliesNoShrink` puts in `obj\Debug\android\assets`.

I updated a test to check a `Release` build with `$(PublishTrimmed)`
set to `false`. This should now be the same as setting
`$(AndroidLinkMode)` to `None`. We will likely need to support both
sets of linker properties.