[ci] Install .NET 5 stable and update VS
All build and test jobs have been updated to install both .NET Core
3.1.201 and .NET 5.0.100.  Eventually, we should be able to remove the
3.1.201 installation after we migrate everything targeting .NET Core to
.NET 5.  Windows jobs will now update the most recent Visual Studio 2019
instance on disk to the latest available stable version.

There are currently multiple issues with .NET 5 and `xabuild`.  First
off, the system Mono that we are using contains a version of MSBuild
that is not compatible with .NET 5 targets, which attempt to use a
static property method that does not exist in Mono's MSBuild:

    /Users/builder/azdo/_work/_tool/dotnet/sdk/5.0.100/Sdks/Microsoft.NET.Sdk/targets/Microsoft.NET.TargetFrameworkInference.targets(54,5): error MSB4186: Invalid static method invocation syntax: "[MSBuild]::GetTargetFrameworkIdentifier('$(TargetFramework)')". Method '[MSBuild]::GetTargetFrameworkIdentifier' not found. Static method invocation should be of the form: $([FullTypeName]::Method()), e.g. $([System.IO.Path]::Combine(`a`, `b`)). Check that all parameters are defined, are of the correct type, and are specified in the right order.

Once this issue is resolved (or in the case of our Windows builds), we
hit the second issue:

    C:\Program Files\dotnet\sdk\5.0.100\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.TargetFrameworkInference.targets(54,5): error MSB4184: The expression "[MSBuild]::GetTargetFrameworkIdentifier(netstandard2.0)" cannot be evaluated. MSB0001: Internal MSBuild Error: A required NuGet assembly was not found. Expected Path: C:\Users\Peter\AppData\Local\Temp\ktbpeg2z.b31

If we decide that xabuild should support .NET 5+ in the future, we'll
need to copy a version of `NuGet.Frameworks.dll` next to
`MSBuildExeTempPath` to work around this issue.