[ci] Restore custom use-dot-net task
Context: https://github.com/xamarin/xamarin-android/pull/5293

PR #5293 introduced an incompatibility with xabuild, and it appears to
be causing issues with other stable builds.  The [UseDotNet][1] task has
a caching feature to improve "installation" times for certain versions,
however this can cause issues for our non-hosted build machines.

    /Users/builder/azdo/_work/_tool/dotnet/sdk/5.0.100/Sdks/Microsoft.NET.Sdk/targets/Microsoft.NET.TargetFrameworkInference.targets(54,5): error MSB4186: Invalid static method invocation syntax: "[MSBuild]::GetTargetFrameworkIdentifier('$(TargetFramework)')". Method '[MSBuild]::GetTargetFrameworkIdentifier' not found. Static method invocation should be of the form: $([FullTypeName]::Method()), e.g. $([System.IO.Path]::Combine(`a`, `b`)). Check that all parameters are defined, are of the correct type, and are specified in the right order. [/Users/builder/azdo/_work/1/s/xamarin-android/external/Java.Interop/src/Java.Interop.GenericMarshaler/Java.Interop.GenericMarshaler.csproj]

We should continue to use the custom `use-dot-net` task that Jonathan
Peppers added for us during earlier .NET 5 preview work.  This template
was removed in PR #5191 but in hindsight it should not have been.

[1]: https://docs.microsoft.com/en-us/azure/devops/pipelines/tasks/tool/dotnet-core-tool-installer?view=azure-devops