[One .NET] Add Microsoft.Android-Tests unit tests (#5250)
A new version of Mono.Android-Tests has been added to run against the
.NET 6 SDK:

   * `tests/Mono.Android-Tests/Runtime-Microsoft.Android.Sdk/Mono.Android.NET-Tests.csproj`

To support this, new .NET `.csproj` files have been added alongside
all of the existing legacy Xamarin.Android unit test library projects.
These projects target `net5.0-android` and build against our
.NET 6 SDK (6f761ea8):

  * `src/Mono.Android/Test/Java.Interop-Tests/Java.Interop-Tests.NET.csproj`
  * `src/Mono.Android/Test/Mono.Android-Test.Library/Mono.Android-Test.Library.NET.csproj`
  * `src/Xamarin.Android.NUnitLite/Xamarin.Android.NUnitLite.NET.csproj`
  * `tests/TestRunner.Core/TestRunner.Core.NET.csproj`
  * `tests/TestRunner.NUnit/TestRunner.NUnit.NET.csproj`

The `_CreateAar` target was being skipped for the new
`Java.Interop-Tests.NET` project, as it had no inputs.  As a result,
the `Mono.Android-Test-classes.jar` file required by these tests was
not being included in the application project.  To fix this, I've added
`@(AndroidJavaLibrary)` to `@(_CreateAarInputs)` so that these items
will also be considered as inputs for this target/task.

All tests which have been ignored are tied back to one of the following issues:

  * Mono.Data.Sqlite is not supported.
  * Secondary AppDomains are not supported.
  * Missing crypto/networking support in dotnet/runtime.
  * https://github.com/dotnet/runtime/issues/35836
  * https://github.com/dotnet/runtime/issues/44358
  * https://github.com/xamarin/xamarin-android/issues/5265
  * https://github.com/xamarin/xamarin-android/issues/5275