[xaprepare] Fix for newer homebrew on macOs
The latest version of homebrew introduced a breaking change in the
way it handles certain combinations of its command-line arguments.
Namely, what affects us, is the combination `--versions -1` to
limit the number of package's versions to a single line. Right
now it is not allowed and causes the following failure:

    Error: Invalid usage: Options --versions and -1 are mutually exclusive.
     [MISSING]
    Checking automake                                   Usage: brew list, ls [options] [formula|cask]

Instead of passing `-1`, do the processing ourselves.