[macOS] More checks for Apple silicon support
The ARM64 build of our native runtime code fails with the following
message when Xcode with Apple silicon support is installed:

    ld : warning : ignoring file libmonosgen-2.0.dylib, building for macOS-arm64 but attempting to link with file built for macOS-x86_64
      Undefined symbols for architecture arm64:
        "_mono_add_internal_call", referenced from:
    ...
    ld: symbol(s) not found for architecture arm64
    clang : error : linker command failed with exit code 1 (use -v to see invocation)

This is caused by the fact that the Mono archive we use does not
contain Mono binaries which include both the `x86_64` and the
`arm64` architectures.

Add a check on `libmonosgen-2.0.dylib` to see if it's a fat binary
with both required architectures and disable ARM64 if not, even if
the Xcode toolchain supports Apple silicon.