[l10n] Sort .resx file (#5296)
Two of the entries in the `.resx` file were out of sort order based
on the `//data/@name` attributes.  Because the file was out of order,
the Visual Studio `.resx` editor would preserve that custom order by
appending all new entries to the bottom of the file instead of
placing them in sorted positions.

Correct the order so that new entries added using the Visual Studio
`.resx` editor will be placed automatically in sorted positions.