[Xamarin.Andrid.Build.Tests] Add MultiUser unit tests
This commit adds support for running multi user tests on device.
It makes use of various adb commands to create test users and
remove them.

`shell pm create-user --guest {name}` will create a guest user of a
specific name.

`shell pm remove-user --guest {userId}` will remove a userId.

The userId can be retrieved via `shell pm list users`. This command
will return results as follows

    Users:
	    UserInfo{0:Owner:c13} running
	    UserInfo{10:guest1:414}

We can then use a RegEx to pull out the required `userId` for a specific
user.

Finally `shell am switch-user {userId}` can be used to switch between
users on the device.

Emulators seem to create the main user with the name "Owner" so that should
be the default name we use for the main user. In the unit tests if we
cannot find a user named "Owner" when looking for "Owner" we will default
to a `userId` of `0`. This always means the default user. This is to handle
the case where we are testing locally on devices where the main user is
not called "Owner".