[Xamarin.Android.Build.Tasks] fix incremental _ManifestMerger target
When testing a .NET 6 project using AndroidX and
`$(AndroidManifestMerger)` = `manifestmerger.jar`, I noticed:

1. I edited the `.csproj` file and did various incremental builds.
2. During `dotnet build`, the `_ManifestMerger` MSBuild target was
   always running--even if there were no changes.

I was able to reproduce the issue in a test, even in "legacy"
Xamarin.Android:

    Building target "_ManifestMerger" completely.
    Input file "obj\Debug\AndroidManifest.xml" is newer than output file "obj\Debug\android\AndroidManifest.xml".

What appears to be happening is:

1. The `Inputs` to `_ManifestMerger` changed.
2. The `Outputs` had identical contents and so
   `MonoAndroidHelper.CopyIfStreamChanged()` did not actually write a
   file.

At this point you are stuck in a place where the `Inputs` will be
perpetually newer than the `Outputs`.

To solve the issue, we can simply call `File.SetLastWriteTimeUtc()`
when `CopyIfStreamChanged()` returns `false` in the `<ManifestMerger/>`
MSBuild task.

This saved ~296ms for incremental builds of a "Hello World" project on
my Windows PC.