[One .NET] move Properties\AndroidManifest.xml to project root
In "legacy" Xamarin.Android, projects default to:

    Properties\AndroidManifest.xml
    Properties\AssemblyInfo.cs
    Foo.csproj

Since `AssemblyInfo.cs` is generally not used in short-form MSBuild
projects, `AndroidManifest.xml` is the only file in the `Properties`
folder.

Let's move the default location for `AndroidManifest.xml` in .NET 6 to
clean things up. Xamarin.iOS also puts the `Info.plist` in the top
directory, so this aligns both platforms better.

To help with migration the old path will continue to work by default
if the file is found:

    <AndroidManifest Condition=" '$(AndroidManifest)' == '' and Exists ('Properties\AndroidManifest.xml') and !Exists ('AndroidManifest.xml') ">Properties\AndroidManifest.xml</AndroidManifest>
    <AndroidManifest Condition=" '$(AndroidManifest)' == '' ">AndroidManifest.xml</AndroidManifest>

We will default to simply `AndroidManifest.xml` otherwise and make
sure we pick `AndroidManifest.xml` if both are found.

I updated `XASdkProject` to use the new path. MSBuild tests that use
`XamarinAndroidApplicationProject` under a `dotnet` context will
continue to use `Properties\AndroidManifest.xml`, so we have test
coverage of both.