[Xamarin.Android.Build.Tasks] Add support for Proguard `mapping.txt` file. (#5304)
Fixes: https://github.com/xamarin/xamarin-android/issues/5186

When uploading a package to the Google Play Store users are seeing
the following warning:

	This App Bundle contains Java/Kotlin code, which might be obfuscated.

To fix this users need to provide a `mapping.txt` file which contains
the Java class mappings from their plain versions to the ones which
are obfuscated.  By default we do not obfuscate the java code but the
warning still appears.

To fix this issue we have a new `$(AndroidProguardMappingFile)` MSBuild
property which defaults to `$(OutputPath)mapping.txt`.  This file will
be produced as part of the build process.

In order for this mapping file to be generated, the following lines
needed to be added to `proguard_xamarin.cfg`:

	-keepattributes SourceFile
	-keepattributes LineNumberTable