[Xamarin.Android.Build.Tasks] Add resource file conflict errors and warnings
Fixes: https://github.com/xamarin/xamarin-android/issues/3702
Context: https://github.com/xamarin/monodroid/commit/1297ecc98dde7b97ab5db771357a93878b3003ad
Context: https://github.com/xamarin/monodroid/commit/0c65bbbb5d7324954aad129b8cc4a892ed87b245
Context: https://bugzilla.xamarin.com/show_bug.cgi?id=28721

There are several ways that resource and asset files can produce
unexpected results with the current way `<AndroidComputeResPaths/>`
works in combination with the `_GenerateAndroidResourceDir` and
`_GenerateAndroidAssetsDir` targets.

One example is if a project references two Android layout resource files
that only differ by their file extensions, like:

  * `Resources\layout\layout1.axml`
  * `Resources\layout\layout1.xml`

To give projects the flexibility to use either extension, the
`<AndroidComputeResPaths/>` intentionally replaces `.axml` with `.xml`
when computing the intermediate output location.  But this means that
both files end up with the same intermediate output location:

	obj\Debug\100\res\layout\layout1.xml

Consequently, the`<CopyIfChanged/>` task in the
`_GenerateAndroidResourceDir` target copies both files to that single
output location, one after the other, overwriting the contents of the
first file with the contents of the second file.

As an example of what a user might see in this case, if the first file
defines resource IDs that don't exist in the second file, this can
result in build errors similar to:

	'Resource.Id' does not contain a definition for 'textView1'

The same kind of conflict can also arise on a case-sensitive file system
if two resource files only differ by capitalization.

Similar conflicts can also occur if the two resource or asset items
refer to different on-disk files but have the `%(Link)` or
`%(LogicalName)` metadata that are the same or only differ by file
extension or capitalization.

A less problematic scenario is if two resource or asset items refer to
the *same* on-disk file and have `%(Link)` or `%(LogicalName)` metadata
that only differ by file extension or capitalization.  In that case, the
app won't be missing any data, so it's fine to continue the build with
just a warning rather than an error.  That was the scenario discussed in
the past in [monodroid/0c65bbbb][0] and [Bugzilla bug 28721][1].

To improve the visibility of these complications, adjust
`<AndroidComputeResPaths/>` to emit errors and warnings for these
different scenarios.

Add corresponding tests as well.  Refactor the existing test in
`AndroidComputeResPathsTests` a little along the way so that all the
tests can share some setup code.

The time cost of the new check in `<AndroidComputeResPaths/>` looks
negligible.  I timed the task locally across three trials before and
then three trials after the changes, using a scenario with 10,000 input
file names.

Before with `LowercaseFilenames = true` (for resources):
	860 ms, 856 ms, 843 ms. Average: 853 ms

After with `LowercaseFilenames = true` (for resources):
	823 ms, 851 ms, 845 ms. Average: 840 ms

Before with `LowercaseFilenames = false` (for assets):
	875 ms, 808 ms, 804 ms. Average: 829 ms

After with `LowercaseFilenames = false` (for assets):
	789 ms, 853 ms, 846 ms. Average: 829 ms

[0]: https://github.com/xamarin/monodroid/commit/0c65bbbb5d7324954aad129b8cc4a892ed87b245
[1]: https://bugzilla.xamarin.com/show_bug.cgi?id=28721
[2]: https://github.com/mono/monodevelop/blob/c1e35b5a85a361f7a1dcc6b10b7da105553479e1/main/src/core/MonoDevelop.Core/MonoDevelop.Core/FileService.cs#L136-L149