[One .NET] ignore @(Content) from @(ProjectReference) (#5308)
Fixes: https://github.com/xamarin/xamarin-android/issues/5299

Context: https://github.com/dotnet/sdk/issues/14619

When setting up `Mono.Android.NET-Tests`, we had to work around an
error with duplicate `@(Content)` files:

	error NETSDK1148: Found multiple publish output files with the same relative path: external\Java.Interop\bin\Release\java-interop.jar, \external\Java.Interop\bin\Release-netcoreapp3.1\java-interop.jar.

Android doesn't support the `@(Content)` build action *at all*, and
we emit a `XA0101` warning if it is used. We even have a test for this.

However, it appears that any `@(Content)` from `@(ProjectReference)`
items are currently brought over in .NET 6, as they get added to the
`@(ResolvedFileToPublish)` item group by the dotnet/sdk.

Luckily there appears to be a
[`$(_GetChildProjectCopyToPublishDirectoryItems)`][0] MSBuild property
we can set to skip this behavior.

We can set this by default, as we don't need to compute `@(Content)`
items in referenced projects.  The library project will emit `XA0101`
and that is enough.  I also updated a test that would have triggered
`NETSDK1148` which now works.

I could also see a perf improvement in the `Mono.Android.NET-Tests`
project on Windows:

  * Before:

        GetCopyToPublishDirectoryItems 21ms

  * After:

        GetCopyToPublishDirectoryItems 0ms


[0]: https://github.com/dotnet/sdk/blob/955c0fc7b06e2fa34bacd076ed39f61e4fb61716/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.Publish.targets#L783-L794