[Xamarin.Android.Build.Tasks] Allow users to specify modules for apksets.
Context #4810

When using app bundles users install the application from an `apkset`.
At this time this only contains the base application. However we might
in the future support dynamic features in some fashion. In this case
developers should be able to pick which features they want installed.

This commit adds an `AndroidInstallModules` ItemGroup which will allow
developers to specify which modules to install. By default it will be
empty. This means all modules which are to be installed on first install
will be installed. So there is no change in the current behavior.

This commit also adds support for passing additional arguments to
`bundletool` via the new `$(AndroidBundleToolExtraArgs)` property.