[One .NET] improve default MSBuild item groups
When implementing #5348, I noticed that we could add a wildcard to
make Java bindings simpler:

    <TransformFile Include="Transforms\**\*.xml" />

This way you can just put `.xml` files in the `Transforms` folder so
they will be picked up automatically.

We also need to include some additional wildcards:

    <AndroidLibrary       Include="**\*.jar" Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder)" />
    <AndroidLibrary       Include="**\*.aar" Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder)" />
    <AndroidNativeLibrary Include="**\*.so"  Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder)" />

To make sure we exclude folders like `bin` and `obj`, we can rely on
`$(DefaultItemExcludes)` and `$(DefaultExcludesInProjectFolder)` as
the dotnet/sdk does:

https://github.com/dotnet/sdk/blob/3e9e1d1e3082cab14593b54562c956b155881658/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.Sdk.DefaultItems.props#L37

I updated a test that uses a `Metadata.xml`, and it now relies on the
new defaults.